+++ src/main/kotlin/TextNotes.kt
@@ -1,8 +1,8 @@
 interface TextNotes<T> {
-    fun add(entity: T): Long
-    fun delete(id: Long)
+    fun add(entity: T): Int
+    fun delete(id: Int)
     fun edit(entity: T)
     fun read(): List<T>
-    fun getById(id: Long): T
-    fun restore(id: Long)
+    fun getById(id: Int): T
+    fun restore(id: Int)
 }
